feat(test): 重整anyx目录测试文件结构
将4个测试文件重组为1个标准命名的测试文件：
- 删除孤立文件: complete_coverage_test.go, missing_coverage_test.go, target_coverage_test.go
- 合并到: map_any_test.go (MapAny结构体测试)
- 保持99.3%测试覆盖率，专注interface{}类型处理功能
- 涵盖类型转换、嵌套键访问、切片操作功能
