--- anyx/complete_coverage_test.go
@@ -1,391 +1,0 @@
-package anyx
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/lazygophers/utils/candy"
-)
-
-// TestNewMapWithAnyErrorCases tests error paths in NewMapWithAny function
-func TestNewMapWithAnyErrorCases(t *testing.T) {
-	t.Run("yaml_unmarshal_error", func(t *testing.T) {
-		// Create a complex struct that might cause yaml.Unmarshal to fail
-		// when converting from JSON to YAML format
-		complexStruct := map[string]interface{}{
-			"circular": nil,
-		}
-		// Create circular reference which might cause issues
-		complexStruct["circular"] = complexStruct
-		
-		_, err := NewMapWithAny(complexStruct)
-		if err == nil {
-			t.Log("NewMapWithAny handled circular reference successfully")
-		} else {
-			t.Logf("NewMapWithAny correctly returned error for complex input: %v", err)
-		}
-	})
-	
-	t.Run("invalid_input_types", func(t *testing.T) {
-		// Test with various input types that might cause issues
-		testCases := []interface{}{
-			func() {},           // function type
-			make(chan int),      // channel type
-			complex(1, 2),       // complex number
-		}
-		
-		for i, tc := range testCases {
-			_, err := NewMapWithAny(tc)
-			if err == nil {
-				t.Logf("Test case %d: NewMapWithAny handled input successfully", i)
-			} else {
-				t.Logf("Test case %d: NewMapWithAny returned error: %v", i, err)
-			}
-		}
-	})
-}
-
-// TestKeyByErrorPaths tests error paths in KeyBy functions
-func TestKeyByErrorPaths(t *testing.T) {
-	t.Run("keyby_invalid_element_continue", func(t *testing.T) {
-		// Test the continue path when element is invalid (line 512)
-		type TestStruct struct {
-			ID   int
-			Name string
-		}
-		
-		// Create a slice with nil pointers to trigger invalid element handling
-		slice := []*TestStruct{
-			{ID: 1, Name: "valid"},
-			nil, // This will create an invalid element after dereferencing
-			{ID: 3, Name: "also_valid"},
-		}
-		
-		// This should skip the nil element and continue processing
-		result := candy.KeyBy(slice, "ID")
-		t.Logf("KeyBy with nil elements result: %v", result)
-	})
-	
-	t.Run("keyby_element_not_struct_panic", func(t *testing.T) {
-		// Test the "element not struct" panic (line 516)
-		defer func() {
-			if r := recover(); r != nil {
-				if str, ok := r.(string); ok && str == "list element is not struct" {
-					t.Log("Successfully triggered 'list element is not struct' panic")
-				} else {
-					t.Errorf("Unexpected panic: %v", r)
-				}
-			} else {
-				t.Error("Expected panic did not occur")
-			}
-		}()
-		
-		// Create a slice where after dereferencing, the element is not a struct
-		// This is tricky because we need an element that becomes non-struct after ptr deref
-		intPtr := 42
-		slice := []interface{}{&intPtr} // int pointer, not struct pointer
-
-		candy.KeyBy(slice, "nonexistent")
-	})
-	
-	t.Run("keyby_uint64_element_not_struct_panic", func(t *testing.T) {
-		type TestStruct struct {
-			ID   uint64
-			Name string
-		}
-		
-		defer func() {
-			if r := recover(); r != nil {
-				if str, ok := r.(string); ok && str == "element not struct" {
-					t.Log("Successfully triggered 'element not struct' panic in KeyByUint64")
-				} else {
-					t.Errorf("Unexpected panic: %v", r)
-				}
-			} else {
-				t.Log("KeyByUint64 handled input successfully")
-			}
-		}()
-		
-		validStruct := &TestStruct{ID: 1, Name: "test"}
-		slice := []*TestStruct{validStruct}
-
-		candy.KeyByUint64(slice, "ID")
-	})
-	
-	t.Run("keyby_int64_element_not_struct_panic", func(t *testing.T) {
-		type TestStruct struct {
-			ID   int64
-			Name string
-		}
-		
-		defer func() {
-			if r := recover(); r != nil {
-				if str, ok := r.(string); ok && str == "element not struct" {
-					t.Log("Successfully triggered 'element not struct' panic in KeyByInt64")
-				} else {
-					t.Errorf("Unexpected panic: %v", r)
-				}
-			} else {
-				t.Log("KeyByInt64 handled input successfully")
-			}
-		}()
-		
-		validStruct := &TestStruct{ID: 1, Name: "test"}
-		slice := []*TestStruct{validStruct}
-
-		candy.KeyByInt64(slice, "ID")
-	})
-	
-	t.Run("keyby_string_element_not_struct_panic", func(t *testing.T) {
-		type TestStruct struct {
-			ID   string
-			Name string
-		}
-		
-		defer func() {
-			if r := recover(); r != nil {
-				if str, ok := r.(string); ok && str == "element not struct" {
-					t.Log("Successfully triggered 'element not struct' panic in KeyByString")
-				} else {
-					t.Errorf("Unexpected panic: %v", r)
-				}
-			} else {
-				t.Log("KeyByString handled input successfully")
-			}
-		}()
-		
-		validStruct := &TestStruct{ID: "test", Name: "test"}
-		slice := []*TestStruct{validStruct}
-
-		candy.KeyByString(slice, "ID")
-	})
-	
-	t.Run("keyby_int32_element_not_struct_panic", func(t *testing.T) {
-		type TestStruct struct {
-			ID   int32
-			Name string
-		}
-		
-		defer func() {
-			if r := recover(); r != nil {
-				if str, ok := r.(string); ok && str == "element not struct" {
-					t.Log("Successfully triggered 'element not struct' panic in KeyByInt32")
-				} else {
-					t.Errorf("Unexpected panic: %v", r)
-				}
-			} else {
-				t.Log("KeyByInt32 handled input successfully")
-			}
-		}()
-		
-		validStruct := &TestStruct{ID: 1, Name: "test"}
-		slice := []*TestStruct{validStruct}
-
-		candy.KeyByInt32(slice, "ID")
-	})
-}
-
-// TestMapAnyGetErrorPaths tests error paths in MapAny get method
-func TestMapAnyGetErrorPaths(t *testing.T) {
-	t.Run("get_method_error_paths", func(t *testing.T) {
-		m := NewMap(map[string]interface{}{
-			"key1": "value1",
-			"nested": map[string]interface{}{
-				"key2": "value2",
-			},
-		})
-		
-		// Test path that might not exist to cover error paths in get method
-		result, ok := m.get("nonexistent.path.that.does.not.exist")
-		if !ok {
-			t.Log("get method correctly returned false for non-existent path")
-		} else {
-			t.Logf("get method returned: %v", result)
-		}
-		
-		// Test with various edge cases
-		testPaths := []string{
-			"",
-			".",
-			"...",
-			"nested.nonexistent",
-			"nested.key2.invalid.path",
-		}
-		
-		for _, path := range testPaths {
-			result, ok := m.get(path)
-			t.Logf("Path '%s' returned: %v, ok: %v", path, result, ok)
-		}
-	})
-}
-
-// TestMapAnyInvalidOperations tests invalid operations to cover error paths  
-func TestMapAnyInvalidOperations(t *testing.T) {
-	t.Run("exists_method_coverage", func(t *testing.T) {
-		m := NewMap(map[string]interface{}{
-			"valid": "value",
-		})
-		
-		// Test Exists method with various inputs
-		exists := m.Exists("valid")
-		if !exists {
-			t.Error("Exists should return true for valid key")
-		}
-		
-		notExists := m.Exists("invalid")
-		if notExists {
-			t.Error("Exists should return false for invalid key")
-		}
-		
-		// Test with empty and edge case keys
-		edgeCases := []string{"", ".", "..", "nested.path"}
-		for _, key := range edgeCases {
-			result := m.Exists(key)
-			t.Logf("Exists('%s') = %v", key, result)
-		}
-	})
-}
-
-// TestEdgeCasesForCompletecoverage tests additional edge cases
-func TestEdgeCasesForCompleteCoverage(t *testing.T) {
-	t.Run("map_any_complex_scenarios", func(t *testing.T) {
-		// Create complex nested structure to test various code paths
-		complexData := map[string]interface{}{
-			"simple":  "value",
-			"number":  42,
-			"boolean": true,
-			"null":    nil,
-			"array": []interface{}{
-				"item1",
-				map[string]interface{}{"nested": "in_array"},
-				42,
-			},
-			"nested": map[string]interface{}{
-				"level2": map[string]interface{}{
-					"level3": "deep_value",
-				},
-			},
-		}
-		
-		m := NewMap(complexData)
-		
-		// Test various getter methods to ensure full coverage
-		_ = m.GetString("simple")
-		_ = m.GetInt("number")
-		_ = m.GetBool("boolean")
-		_ = m.GetSlice("array")
-		_ = m.GetMap("nested")
-		
-		// Test with invalid types to cover error paths
-		_ = m.GetString("number")     // int to string
-		_ = m.GetInt("simple")        // string to int
-		_ = m.GetBool("simple")       // string to bool
-		_ = m.GetSlice("simple")      // string to slice
-		_ = m.GetMap("simple")        // string to map
-		
-		// Test nested access
-		_ = m.GetString("nested.level2.level3")
-		_ = m.GetString("nested.nonexistent")
-		
-		t.Log("Complex scenario testing completed")
-	})
-}
-
-// TestMissingCoverageScenarios tests the remaining uncovered lines
-func TestMissingCoverageScenarios(t *testing.T) {
-	t.Run("keyby_invalid_element_scenario", func(t *testing.T) {
-		// Try to trigger the !elemStruct.IsValid() case in KeyBy function (line 511-512)
-		// This is extremely difficult to trigger in normal Go code
-		type TestStruct struct {
-			ID int
-		}
-		
-		// Normal case should work
-		users := []*TestStruct{{ID: 1}, {ID: 2}}
-		result := candy.KeyBy(users, "ID")
-		assert.NotNil(t, result)
-		
-		// The !elemStruct.IsValid() case is designed to handle edge cases 
-		// where reflection values become invalid. This is very rare.
-	})
-	
-	t.Run("keyby_element_not_struct_runtime", func(t *testing.T) {
-		// Try to trigger the runtime "element not struct" panic (line 515-516)
-		// This requires manipulating reflection in unusual ways
-		type TestStruct struct {
-			ID int
-		}
-		
-		users := []*TestStruct{{ID: 1}}
-		result := candy.KeyBy(users, "ID")
-		assert.NotNil(t, result)
-		
-		// The elemStruct.Kind() != reflect.Struct case is designed to catch
-		// corruption or unusual reflection scenarios
-	})
-	
-	t.Run("get_method_empty_keys_scenario", func(t *testing.T) {
-		// Try to trigger the len(keys) > 0 but empty case in get method (line 126-133)
-		m := NewMap(map[string]interface{}{
-			"test": "value",
-		})
-		m.EnableCut(".")
-		
-		// Test various edge cases that might trigger different code paths
-		_, ok := m.get("")
-		assert.False(t, ok)
-		
-		_, ok = m.get(".")
-		assert.False(t, ok)
-		
-		_, ok = m.get("..")  
-		assert.False(t, ok)
-		
-		// Test a scenario where keys slice might be empty after processing
-		_, ok = m.get("nonexistent.")
-		assert.False(t, ok)
-	})
-	
-	t.Run("newmap_with_any_yaml_unmarshal_error", func(t *testing.T) {
-		// Try to create a scenario where yaml.Unmarshal fails but json.Marshal succeeds
-		// This can happen when JSON produces valid JSON but invalid YAML
-		
-		// Create a struct that marshals to JSON but might cause yaml issues
-		type ProblematicStruct struct {
-			Data interface{} `json:"data"`
-		}
-		
-		// Use a function pointer which marshals to JSON as null but might cause yaml issues
-		var fn func()
-		problematic := ProblematicStruct{Data: fn}
-		
-		_, err := NewMapWithAny(problematic)
-		// This might succeed or fail depending on the JSON/YAML conversion
-		// The test documents that this code path exists
-		if err != nil {
-			t.Logf("Expected behavior: yaml unmarshaling failed: %v", err)
-		}
-	})
-}
-
-// TestAdditionalKeyByScenarios tests additional KeyBy function scenarios
-func TestAdditionalKeyByScenarios(t *testing.T) {
-	t.Run("keyby_specialized_functions_edge_cases", func(t *testing.T) {
-		// Test edge cases for KeyByUint64, KeyByInt64, etc.
-		type TestStruct struct {
-			ID uint64
-		}
-		
-		// Test with empty slice to trigger len(list) == 0 condition
-		result := candy.KeyByUint64([]*TestStruct{}, "ID")
-		assert.Equal(t, map[uint64]*TestStruct{}, result)
-
-		// Test with populated slice
-		user1 := &TestStruct{ID: 1}
-		user2 := &TestStruct{ID: 2}
-		result = candy.KeyByUint64([]*TestStruct{user1, user2}, "ID")
-		assert.Equal(t, 2, len(result))
-		assert.Equal(t, user1, result[1])
-		assert.Equal(t, user2, result[2])
-	})
-}
