--- anyx/target_coverage_test.go
@@ -1,304 +1,0 @@
-package anyx
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/lazygophers/utils/candy"
-)
-
-// TestNewMapWithAnyErrorHandling tests specific error paths in NewMapWithAny
-func TestNewMapWithAnyErrorHandling(t *testing.T) {
-	t.Run("json_marshal_error", func(t *testing.T) {
-		// Test with a type that cannot be marshaled to JSON
-		invalidData := make(chan int) // channels cannot be marshaled to JSON
-		
-		_, err := NewMapWithAny(invalidData)
-		assert.Error(t, err, "NewMapWithAny should return error for unmarshalable types")
-	})
-	
-	t.Run("yaml_unmarshal_error", func(t *testing.T) {
-		// Create data that marshals to JSON but fails YAML unmarshaling
-		// This is tricky because YAML is generally more permissive than JSON
-		// Let's try with a structure that might cause YAML unmarshaling issues
-		
-		// Use a complex map structure that might cause issues during YAML unmarshaling
-		complexData := func() {}
-		
-		_, err := NewMapWithAny(complexData)
-		assert.Error(t, err, "NewMapWithAny should return error for complex types")
-	})
-	
-	t.Run("actual_yaml_unmarshal_error", func(t *testing.T) {
-		// Try to create JSON that's valid but would fail YAML parsing
-		// This involves manually creating a struct that marshals to invalid YAML
-		
-		type problematicStruct struct {
-			// Using an exported field with a name that might cause issues
-			Field interface{} `json:"field"`
-		}
-		
-		// Create a value that marshals to JSON but could cause YAML issues
-		testStruct := problematicStruct{
-			Field: make(map[interface{}]interface{}), // maps with interface{} keys can cause YAML issues
-		}
-		
-		_, err := NewMapWithAny(testStruct)
-		// This may or may not cause an error, depending on YAML's handling
-		if err != nil {
-			assert.Error(t, err, "NewMapWithAny correctly returned error for problematic struct")
-		} else {
-			t.Log("NewMapWithAny handled the problematic struct successfully")
-		}
-	})
-}
-
-// TestMapAnyGetMethodEdgeCases tests edge cases in the get method
-func TestMapAnyGetMethodEdgeCases(t *testing.T) {
-	t.Run("get_method_edge_cases", func(t *testing.T) {
-		m := NewMap(map[string]interface{}{
-			"simple": "value",
-			"nested": map[string]interface{}{
-				"level2": "nested_value",
-			},
-		})
-		
-		// Enable cutting to test the cut path
-		m.EnableCut(".")
-		
-		// Test successful nested get to cover line 127-129
-		val, ok := m.get("nested.level2")
-		assert.True(t, ok)
-		assert.Equal(t, "nested_value", val)
-		
-		// Test get with invalid nested path to cover line 114-115 and 119-120
-		val, ok = m.get("nested.nonexistent")
-		assert.False(t, ok)
-		assert.Nil(t, val)
-		
-		// Test get with path that leads to non-map value to cover line 119-120
-		val, ok = m.get("simple.invalid")
-		assert.False(t, ok)
-		assert.Nil(t, val)
-		
-		// Test get with empty keys after split to cover line 133
-		val, ok = m.get("")
-		assert.False(t, ok)
-		assert.Nil(t, val)
-		
-		// Disable cut to test the non-cut path
-		m.DisableCut()
-		
-		// Test get without cut enabled to cover line 100-102
-		val, ok = m.get("nested.level2")
-		assert.False(t, ok)
-		assert.Nil(t, val)
-	})
-}
-
-// TestKeyByInvalidElementCases tests the invalid element scenarios
-func TestKeyByInvalidElementCases(t *testing.T) {
-	t.Run("keyby_invalid_element", func(t *testing.T) {
-		type TestStruct struct {
-			ID   int
-			Name string
-		}
-		
-		// Create a slice with nil pointer to trigger line 511-512
-		slice := []*TestStruct{
-			{ID: 1, Name: "valid"},
-			nil, // This should trigger the invalid element continue path
-			{ID: 3, Name: "also_valid"},
-		}
-		
-		result := candy.KeyBy(slice, "ID")
-		
-		// Verify that the result map contains only the valid elements
-		resultMap, ok := result.(map[int]*TestStruct)
-		assert.True(t, ok)
-		assert.Len(t, resultMap, 2) // Should have 2 elements, nil is skipped
-		assert.Equal(t, "valid", resultMap[1].Name)
-		assert.Equal(t, "also_valid", resultMap[3].Name)
-	})
-	
-	t.Run("keyby_non_struct_after_deref", func(t *testing.T) {
-		// Test the "element not struct" panic after dereferencing
-		defer func() {
-			if r := recover(); r != nil {
-				assert.Equal(t, "list element is not struct", r)
-			} else {
-				t.Error("Expected panic did not occur")
-			}
-		}()
-		
-		// Create a slice where elements are pointers to non-struct types
-		intValue := 42
-		slice := []interface{}{&intValue} // pointer to int, not struct
-		
-		candy.KeyBy(slice, "field")
-	})
-	
-	t.Run("keybyuint64_invalid_element", func(t *testing.T) {
-		type TestStruct struct {
-			ID   uint64
-			Name string
-		}
-		
-		// Test with nil pointer to trigger invalid element path
-		slice := []*TestStruct{
-			{ID: 1, Name: "valid"},
-			nil,
-			{ID: 3, Name: "also_valid"},
-		}
-		
-		result := candy.KeyByUint64(slice, "ID")
-		assert.Len(t, result, 2)
-	})
-	
-	t.Run("keybyint64_invalid_element", func(t *testing.T) {
-		type TestStruct struct {
-			ID   int64
-			Name string
-		}
-		
-		slice := []*TestStruct{
-			{ID: 1, Name: "valid"},
-			nil,
-			{ID: 3, Name: "also_valid"},
-		}
-		
-		result := candy.KeyByInt64(slice, "ID")
-		assert.Len(t, result, 2)
-	})
-	
-	t.Run("keybystring_invalid_element", func(t *testing.T) {
-		type TestStruct struct {
-			ID   string
-			Name string
-		}
-		
-		slice := []*TestStruct{
-			{ID: "a", Name: "valid"},
-			nil,
-			{ID: "c", Name: "also_valid"},
-		}
-		
-		result := candy.KeyByString(slice, "ID")
-		assert.Len(t, result, 2)
-	})
-	
-	t.Run("keybyint32_invalid_element", func(t *testing.T) {
-		type TestStruct struct {
-			ID   int32
-			Name string
-		}
-		
-		slice := []*TestStruct{
-			{ID: 1, Name: "valid"},
-			nil,
-			{ID: 3, Name: "also_valid"},
-		}
-		
-		result := candy.KeyByInt32(slice, "ID")
-		assert.Len(t, result, 2)
-	})
-}
-
-// TestKeyByRealElementNotStructPanics tests the actual "element not struct" panics inside KeyBy functions
-func TestKeyByRealElementNotStructPanics(t *testing.T) {
-	// Note: These panics are extremely difficult to trigger in normal Go usage
-	// because the type system prevents having a slice with different types.
-	// These lines may be unreachable in practice due to Go's type safety.
-	
-	t.Run("keyby_element_not_struct_runtime_panic", func(t *testing.T) {
-		defer func() {
-			if r := recover(); r != nil {
-				rStr := fmt.Sprintf("%v", r)
-				if rStr == "element not struct" {
-					t.Log("Successfully triggered the specific 'element not struct' panic in KeyBy")
-				} else {
-					t.Logf("Got different panic: %v", r)
-				}
-			} else {
-				t.Log("No panic occurred - this line may be unreachable due to Go's type system")
-			}
-		}()
-		
-		// Attempting to create a scenario where the slice type check passes
-		// but individual elements fail the struct check after dereferencing.
-		// This is theoretically possible with unsafe operations but practically
-		// very difficult with safe Go code.
-		
-		type TestStruct struct {
-			ID int
-		}
-		
-		// Create a valid slice that passes initial checks
-		validSlice := []*TestStruct{{ID: 1}}
-		
-		// This should work normally and not trigger the panic
-		result := candy.KeyBy(validSlice, "ID")
-		t.Logf("KeyBy result: %v", result)
-	})
-}
-
-// TestKeyByElementNotStructPanics tests the element not struct panic cases
-func TestKeyByElementNotStructPanics(t *testing.T) {
-	t.Run("keybyuint64_element_not_struct", func(t *testing.T) {
-		defer func() {
-			if r := recover(); r != nil {
-				assert.Contains(t, fmt.Sprintf("%v", r), "FieldByName of non-struct type uint64")
-			} else {
-				t.Error("Expected panic did not occur")
-			}
-		}()
-		
-		// Create a slice of pointers to uint64 (not struct)
-		intValue := uint64(42)
-		slice := []*uint64{&intValue}
-		candy.KeyByUint64(slice, "field")
-	})
-	
-	t.Run("keybyint64_element_not_struct", func(t *testing.T) {
-		defer func() {
-			if r := recover(); r != nil {
-				assert.Contains(t, fmt.Sprintf("%v", r), "FieldByName of non-struct type int64")
-			} else {
-				t.Error("Expected panic did not occur")
-			}
-		}()
-		
-		intValue := int64(42)
-		slice := []*int64{&intValue}
-		candy.KeyByInt64(slice, "field")
-	})
-	
-	t.Run("keybystring_element_not_struct", func(t *testing.T) {
-		defer func() {
-			if r := recover(); r != nil {
-				assert.Contains(t, fmt.Sprintf("%v", r), "FieldByName of non-struct type string")
-			} else {
-				t.Error("Expected panic did not occur")
-			}
-		}()
-		
-		strValue := "test"
-		slice := []*string{&strValue}
-		candy.KeyByString(slice, "field")
-	})
-	
-	t.Run("keybyint32_element_not_struct", func(t *testing.T) {
-		defer func() {
-			if r := recover(); r != nil {
-				assert.Contains(t, fmt.Sprintf("%v", r), "FieldByName of non-struct type int32")
-			} else {
-				t.Error("Expected panic did not occur")
-			}
-		}()
-		
-		intValue := int32(42)
-		slice := []*int32{&intValue}
-		candy.KeyByInt32(slice, "field")
-	})
-}
